Move order scripting classes into separate file
